% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_clusters.R
\name{add_clusters}
\alias{add_clusters}
\title{add_clusters}
\usage{
add_clusters(.data, ..., n_clusters = 4, cluster_name = "cluster")
}
\arguments{
\item{.data}{dataframe}

\item{...}{columns to cluster}

\item{n_clusters}{integer}

\item{cluster_name}{column name}
}
\value{
data frame
}
\description{
add_clusters
}
