% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bins.R
\name{bin_cols}
\alias{bin_cols}
\title{Make Bins}
\usage{
bin_cols(
  .data,
  col,
  n_bins = 10,
  bin_type = c("frequency", "width", "value", "kmeans", "xgboost", "woe", "logreg",
    "mdlp"),
  ...,
  target = NULL,
  pretty_labels = F,
  seed = 1,
  method = "mdlp"
)
}
\arguments{
\item{.data}{a data frame}

\item{col}{a column, vector of columns, or tidyselect}

\item{n_bins}{number of bins}

\item{bin_type}{method to make bins}

\item{...}{params to be passed to selected binning method}

\item{target}{unquoted column for supervised binning}

\item{pretty_labels}{logical. If T returns interval label rather than integer rank}

\item{seed}{seed for stochastic binning (xgboost)}

\item{method}{method for bun mdlp}
}
\value{
a data frame
}
\description{
Make bins in a tidy fashion. Adds a column to your data frame containing the integer codes of the specified bins of a certain column.
Bin types are input as logical statements. Multiple bin types can be set to `TRUE`, creating multiple types of bins for one column.
Specifying multiple columns is only intended for supervised binning, so mutliple columns can be simultaneously binned
optimally with respect to a target variable.
}
