% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_equal_value.R
\name{bin_equal_value}
\alias{bin_equal_value}
\title{bin equal value}
\usage{
bin_equal_value(mdb, col, n_bins = 10)
}
\arguments{
\item{mdb}{dataframe}

\item{col}{a numeric vector}

\item{n_bins}{number of bins}
}
\value{
an integer vector
}
\description{
Bins a numeric column such that each bin contains 10% of the total value (sum) of that column.
Intended for positive numeric vectors that make sense to sum, such as sales. Negative and NAs get treated as 0.
The function never puts two rows with the same value into different bins. Accessed by the "value" method of the
bin_cols function.
}
