% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_bins.R
\name{bin_summary}
\alias{bin_summary}
\title{summarize bins}
\usage{
bin_summary(mdb, ...)
}
\arguments{
\item{mdb}{dataframe output from bin_cols}

\item{...}{optional tidyselect specification for specific cols}
}
\value{
a tibble
}
\description{
Returns a summary of all bins created by `bin_cols` in a data frame. Takes no arguments other than the data frame
but relies on regular expressions based of the `bin_cols` output in order to identify the corresponding columns.
}
\examples{

iris \%>\%
bin_cols(Sepal.Width) \%>\%
bin_summary()
}
