% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/five_number_summary.R
\name{five_number_summary}
\alias{five_number_summary}
\title{five number summary}
\usage{
five_number_summary(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a tibble
}
\description{
The five number summary of a numeric vector you would get from `summary` but returned with a tidy output.
}
\examples{

iris$Petal.Width \%>\%
five_number_summary()
}
