% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_summary.R
\name{make_labels}
\alias{make_labels}
\title{make labels}
\usage{
make_labels(mdb, original_col, bucket_col)
}
\arguments{
\item{mdb}{dataframe}

\item{original_col}{original col}

\item{bucket_col}{bucket col}
}
\value{
dataframe
}
\description{
make labels
}
\keyword{internal}
