% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_na}
\alias{make_na}
\title{Make NAs}
\usage{
make_na(.data, ..., vec = c("-", "", " ", "null"))
}
\arguments{
\item{.data}{data frame}

\item{...}{tidyselect specification}

\item{vec}{vector of possible elements to replace with NA}
}
\value{
data frame
}
\description{
Set elements to NA values using tidyselect specification.
Don't use this function on columns of different modes at once.
Defaults to choosing all character columns.
}
