% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_summary.R
\name{numeric_summary}
\alias{numeric_summary}
\title{numeric summary}
\usage{
numeric_summary(mdb, original_col, bucket_col)
}
\arguments{
\item{mdb}{a data frame}

\item{original_col}{original numeric column}

\item{bucket_col}{columns of bins}
}
\value{
a tibble
}
\description{
This function summarizes an arbitrary bin column, with respect to its original column. Can be used to summarize
bins created from any package, or any arbitrary categorical column paired with a numeric column.
}
\examples{

iris \%>\%
numeric_summary(original_col = Sepal.Length, bucket_col = Species)

}
