% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{oner_wrapper}
\alias{oner_wrapper}
\title{one_wrapper}
\usage{
oner_wrapper(
  bin_cols,
  .data,
  abbv,
  bin_method,
  n_bins = n_bins,
  pretty_labels = pretty_labels
)
}
\arguments{
\item{bin_cols}{cols}

\item{.data}{dataframe}

\item{abbv}{char}

\item{bin_method}{char. bin method.}

\item{n_bins}{integer. number of bins}

\item{pretty_labels}{pretty_labels}
}
\value{
output
}
\description{
one_wrapper
}
\keyword{internal}
