% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bins.R
\name{make_bins}
\alias{make_bins}
\title{Make Bins}
\usage{
make_bins(
  .data,
  col,
  n_bins = 10,
  bin_equal_frequency = F,
  bin_equal_width = F,
  bin_equal_value = F,
  bin_kmeans = F,
  bin_xgboost = F,
  ...,
  target = NULL,
  seed = 1
)
}
\arguments{
\item{col}{a column, vector of columns, or tidyselect}

\item{n_bins}{number of bins}

\item{bin_equal_frequency}{logical, bin type equal frequency (quantile, ntile)}

\item{bin_equal_width}{logical, bin type equal width (cut)}

\item{bin_equal_value}{logical, bin type equal value}

\item{bin_kmeans}{logical, bin type kmeans}

\item{bin_xgboost}{logical, supervised binning with xgboost. target must be specified}

\item{...}{params to be passed to selected binning method}

\item{target}{unquoted column for supervised binning}

\item{seed}{seed for stochastic binning (xgboost)}
}
\value{
a tibble
}
\description{
Make bins in a tidy fashion. Adds a column to your data frame containing the integer codes of the specified bins of a certain column.
Bin types are input as logical statements. Multiple bin types can be set to `TRUE`, creating multiple types of bins for one column.
Specifying multiple columns is only intended for supervised binning, so mutliple columns can be simultaneously binned
optimally with respect to a target variable.
}
